% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid.R
\name{euclid}
\alias{euclid}
\title{Euclidean distance function}
\usage{
euclid(p1, p2)
}
\arguments{
\item{p1}{First point or data matrix}

\item{p2}{Second point or data matrix}
}
\value{
The euclidean distance between two points
}
\description{
This function calculates the euclidean distance between two points or two data frames/matrices.
}
\examples{
euclid(p1=X.mat,p2=centers)

euclid(p1=centers,p2=centers2)


}
