% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLUST.R
\name{HCLUST}
\alias{HCLUST}
\title{HCLUST: Hierarchical clustering function}
\usage{
HCLUST(data.matrix, K)
}
\arguments{
\item{data.matrix}{Data matrix with data observations to cluster}

\item{K}{Number of clusters}
}
\value{
Vector of clusters assignments.
}
\description{
This function performs hierarchical clustering for a group of data points for a specified number of clusters K.
}
\examples{
HCLUST(data.matrix=X.mat,K=5)

HCLUST(data.matrix=X.mat,K=10)

}
